+++ NEWS.md
@@ -12,6 +12,7 @@
 - Add the `listify.what` argument to `fuzz()` so that each input in `what`
   is also tested in its listified version.
 - Provide the S3 method for `length()`.
+- Add a column with the inputs tested to the data frame produced by summary().
 
 # CBTF 0.2.0 (2025-07-16)